refactor: replace HeaderList and CookieList with createPropertyList for request and response headers
This commit removes the HeaderList and CookieList classes, replacing them with a unified createPropertyList approach. The new structure utilizes RequestHeaderStore and CookieJarStore for managing headers and cookies, respectively. Additionally, new idx(index) hints have been added to the STATIC_API_HINTS for both request and response headers, enhancing the autocomplete functionality. This refactor streamlines the codebase and improves the handling of headers and cookies in scripts.
