+++ packages/bruno-app/src/utils/codemirror/autocomplete.js
@@ -41,6 +41,7 @@ const STATIC_API_HINTS = {
     'req.headerList.get(name)',
     'req.headerList.one(name)',
     'req.headerList.all()',
+    'req.headerList.idx(index)',
     'req.headerList.count()',
     'req.headerList.has(name)',
     'req.headerList.has(name, value)',
@@ -85,6 +86,7 @@ const STATIC_API_HINTS = {
     'res.headerList.get(name)',
     'res.headerList.one(name)',
     'res.headerList.all()',
+    'res.headerList.idx(index)',
     'res.headerList.count()',
     'res.headerList.has(name)',
     'res.headerList.has(name, value)',
