+++ packages/bruno-js/src/bruno-response.js
@@ -1,6 +1,6 @@
 const { get } = require('@usebruno/query');
 const _ = require('lodash');
-const HeaderList = require('./header-list');
+const { createPropertyList } = require('./property-lists/create-property-list');
 
 class BrunoResponse {
   constructor(res) {
@@ -12,8 +12,8 @@ class BrunoResponse {
     this.responseTime = res ? res.responseTime : null;
     this.url = res?.request ? res.request.protocol + '//' + res.request.host + res.request.path : null;
 
-    // HeaderList in static read-only mode — write methods throw
-    this.headerList = new HeaderList(res, { writable: false });
+    // Static read-only snapshot — write methods throw
+    this.headerList = createPropertyList('res.headerList', { source: res });
 
     // Make the instance callable
     const callable = (...args) => get(this.body, ...args);
