+++ packages/bruno-js/src/property-lists/create-property-list.js
@@ -1,6 +1,9 @@
 const { descriptorFor } = require('./manifest');
 const { assemblePropertyList } = require('./property-list');
 const GrpcMetadataStore = require('./stores/grpc-metadata-store');
+const CookieJarStore = require('./stores/cookie-jar-store');
+const RequestHeaderStore = require('./stores/request-header-store');
+const ArrayStore = require('./stores/array-store');
 
 /**
  * createPropertyList — the single entry point through which every scripting surface
@@ -9,6 +12,9 @@ const GrpcMetadataStore = require('./stores/grpc-metadata-store');
  * to `assemblePropertyList`.
  *
  * Wiring shapes per path:
+ *   'bru.cookies'                  { getUrl, interpolate, createCookieJar, getCookiesForUrl }
+ *   'req.headerList'               { source }  — the raw request config object
+ *   'res.headerList'               { source }  — the response object (may be null)
  *   'bru.grpc.request.metadata'    { readMetadata, writable }
  *   'bru.grpc.response.metadata'   { readMetadata }
  *   'bru.grpc.response.trailers'   { readMetadata }
@@ -18,6 +24,13 @@ const grpcMetadataDispatcher = (descriptor, wiring) =>
   assemblePropertyList(descriptor, new GrpcMetadataStore(wiring.readMetadata), { writable: wiring.writable });
 
 const DISPATCHERS = {
+  'bru.cookies': (descriptor, wiring) => assemblePropertyList(descriptor, new CookieJarStore(wiring)),
+  'req.headerList': (descriptor, wiring) => assemblePropertyList(descriptor, new RequestHeaderStore(wiring.source)),
+  'res.headerList': (descriptor, wiring) => {
+    const rawHeaders = (wiring.source && wiring.source.headers) || {};
+    const items = Object.entries(rawHeaders).map(([key, value]) => ({ key, value }));
+    return assemblePropertyList(descriptor, new ArrayStore(items));
+  },
   'bru.grpc.request.metadata': grpcMetadataDispatcher,
   'bru.grpc.response.metadata': grpcMetadataDispatcher,
   'bru.grpc.response.trailers': grpcMetadataDispatcher
