+++ packages/bruno-js/src/sandbox/quickjs/shims/bru.js
@@ -1,6 +1,7 @@
 const { cleanJson, cleanCircularJson } = require('../../../utils');
 const { marshallToVm } = require('../utils');
 const { createPropertyListBridge } = require('../utils/property-list-bridge');
+const { bridgeMethodSets } = require('../../../property-lists/manifest');
 
 const addBruShimToContext = (vm, bru) => {
   const bruObject = vm.newObject();
@@ -356,10 +357,7 @@ const addBruShimToContext = (vm, bru) => {
   let bruCookiesObject = vm.newObject();
   const { evalCode: cookiesEvalCode } = createPropertyListBridge(vm, bru.cookies, bruCookiesObject, {
     globalPath: 'globalThis.bru.cookies',
-    syncReadMethods: ['get', 'has', 'count', 'indexOf', 'toObject', 'toString'],
-    syncReadObjectMethods: ['one', 'all', 'idx', 'toJSON'],
-    asyncWriteMethods: ['add', 'upsert', 'remove', 'clear', 'delete'],
-    withIterators: true
+    ...bridgeMethodSets('bru.cookies')
   });
 
   const _jarFn = vm.newFunction('_jar', () => {
